requirements.txt
crawl4ai>=0.5.0
pydantic>=2.0.0
pymongo>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
